backend_f1tenth/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
google-genai>=1.0.0
anthropic>=0.40.0
